refactor: merge pvIECTestTemplates2.ts into pvIECTestTemplates.ts (closes #185)
The 10 IEC test equipment templates were split across two files
(parts 1–5 in pvIECTestTemplates.ts, parts 6–10 in pvIECTestTemplates2.ts)
as an editor-performance workaround that no longer applies.

Changes:
- Merge all 10 templates into pvIECTestTemplates.ts
- Export both PV_IEC_TEMPLATES_PART1 and PV_IEC_TEMPLATES_PART2 from the
  single file for backward compatibility
- Update pvIECTestRegistry.ts and pvTestingTemplates.ts to import from
  the merged file
- Delete pvIECTestTemplates2.ts (222 lines removed, zero callers remain)

No logic change — pure file consolidation.
